+++ package.json
@@ -20,7 +20,7 @@
     "bcrypt": "^5.1.1",
     "dotenv": "^16.4.7",
     "mysql2": "^3.12.0",
-    "next": "15.1.3",
+    "next": "^15.1.4",
     "next-auth": "^4.24.11",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
