+++ src/app/(auth)/login/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth/next";
 import { redirect } from "next/navigation";
-import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { authOptions } from "../../../lib/authOptions";
 import LoginForm from "@/components/Forms/LoginForm";
