+++ src/app/user/page.tsx
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth/next";
 import { redirect } from "next/navigation";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "../../lib/authOptions";
 import LogoutButton from "@/components/Logout";
 
 export default async function User(){
